backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.44.0
lime==0.2.0.1
imbalanced-learn==0.11.0
optuna==3.4.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pyarrow==14.0.2
scipy==1.16.3
psutil==7.2.2
aiofiles==25.1.0
fpdf2==2.8.7
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0
websockets==12.0
